==> 4. Benchmark knn_anova.ipynb <==
import pandas as pd
from statsmodels.formula.api import ols
import statsmodels.api as sm

folder_name = "data"
data = pd.read_csv(f'./{folder_name}/all_prec_recall_values.csv')

# Data cleaning: removing unnecessary columns
data_cleaned = data.drop(['Timepoint'], axis=1)

# Renaming the 'Cell Line' column to remove the space
data_cleaned.rename(columns={'Cell Line': 'Cell_Line'}, inplace=True)

# Updating the formula with the corrected column name
formula = 'Precision ~ C(Modality) * C(Perturbation) * C(K)'
formula_recall = 'Recall ~ C(Modality) * C(Perturbation) * C(K)'
# Fit the ANOVA models for Precision and Recall
model_precision = ols(formula, data=data_cleaned).fit()
anova_table_precision = sm.stats.anova_lm(model_precision, typ=2)

model_recall = ols(formula_recall, data=data_cleaned).fit() #[['Modality', 'Perturbation', 'Recall']]
anova_table_recall = sm.stats.anova_lm(model_recall, typ=2)

print("Precision:\n", anova_table_precision)
print("\nRecall:\n", anova_table_recall)
